requirements.txt
streamlit
atproto
google-generativeai
python-dotenv